requirements.txt
pyperclip==1.9.0
PyAutoGUI==0.9.54
google-generativeai==0.8.3
